docs/requirements.txt
sphinxcontrib-bibtex
pandas
openmdao
lxml
ipopo
jupyterlab
matplotlib
plotly
ipywidgets
importlib_resources
ipysheet
tomlkit
wop
